requirements.txt
fastapi
uvicorn
gunicorn
python-multipart
openai
transformers
torch
Pillow
python-dotenv
chardet
huggingface_hub

